Refactored ClassDiagramSVGService, moved service package.
