+++ src/main/java/org/java2uml/java2umlapi/restControllers/ProjectInfoController.java
@@ -7,9 +7,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.java2uml.java2umlapi.fileStorage.ClassDiagramSVGService;
 import org.java2uml.java2umlapi.fileStorage.entity.ProjectInfo;
 import org.java2uml.java2umlapi.fileStorage.repository.ProjectInfoRepository;
+import org.java2uml.java2umlapi.fileStorage.service.ClassDiagramSVGService;
 import org.java2uml.java2umlapi.fileStorage.service.UnzippedFileStorageService;
 import org.java2uml.java2umlapi.lightWeight.service.MethodSignatureToMethodIdMapService;
 import org.java2uml.java2umlapi.modelAssemblers.ProjectInfoAssembler;
